ALL: Phase one of the rebranding of everything to Cabal
The binary will be 'cabalexec' to avoid conflicts. Since the intention
is to make this into a framework, that's fine -- it'll just be the main
executive of the project.

I'm sure I've missed things (not including the logo, which obviously has
to go), and they'll be handled in the future.